+++ auth/auth.go
@@ -40,7 +40,11 @@ func (a *Auth) Init(ab *authboss.Authboss) (err error) {
 
 // LoginGet simply displays the login form
 func (a *Auth) LoginGet(w http.ResponseWriter, r *http.Request) error {
-	return a.Core.Responder.Respond(w, r, http.StatusOK, PageLogin, nil)
+	var data authboss.HTMLData
+	if redir := r.URL.Query().Get(authboss.FormValueRedirect); len(redir) != 0 {
+		data = authboss.HTMLData{authboss.FormValueRedirect: redir}
+	}
+	return a.Core.Responder.Respond(w, r, http.StatusOK, PageLogin, data)
 }
 
 // LoginPost attempts to validate the credentials passed in