+++ authboss.go
@@ -80,16 +80,32 @@ func (a *Authboss) UpdatePassword(ctx context.Context, user AuthableUser, newPas
 	return rmStorer.DelRememberTokens(ctx, user.GetPID())
 }
 
-// Middleware prevents someone from accessing a route they are not allowed to.
-// It allows the user through if they are logged in.
+// Middleware prevents someone from accessing a route that should be
+// only allowed for users who are logged in.
+// It allows the user through if they are logged in (SessionKey).
 //
-// If redirectToLogin is true, the user will be redirected to the login page, otherwise they will
-// get a 404. The redirect goes to: mountPath/login, this means it's expected that the auth module
-// is loaded if this is set to true.
+// If redirectToLogin is true, the user will be redirected to the
+// login page, otherwise they will get a 404.
+// The redirect goes to: mountPath/login, this means it's expected that
+// the auth module is loaded if this is set to true.
 //
-// If allowHalfAuth is true then half-authed users are allowed through, otherwise a half-authed
-// user will not be allowed through.
-func Middleware(ab *Authboss, redirectToLogin bool, allowHalfAuth bool, force2fa bool) func(http.Handler) http.Handler {
+// If forceFullAuth is true then half-authed users (SessionHalfAuth)
+// are not allowed through, otherwise a half-authed user will be allowed through.
+//
+// If force2fa is true, then users must have been logged in
+// with 2fa (Session2FA) otherwise they will not be allowed through.
+func Middleware(ab *Authboss, redirectToLogin bool, forceFullAuth bool, force2fa bool) func(http.Handler) http.Handler {
+	return MountedMiddleware(ab, false, redirectToLogin, forceFullAuth, force2fa)
+}
+
+// MountedMiddleware hides an option from typical users in "mountPathed".
+// Normal routes should never need this only authboss routes (since they
+// are behind mountPath typically). This method is exported only for use
+// by Authboss modules, normal users should use Middleware instead.
+//
+// If mountPathed is true, then before redirecting to a URL it will add
+// the mountpath to the front of it.
+func MountedMiddleware(ab *Authboss, mountPathed, redirectToLogin, forceFullAuth, force2fa bool) func(http.Handler) http.Handler {
 	return func(next http.Handler) http.Handler {
 		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			log := ab.RequestLogger(r)
@@ -98,7 +114,12 @@ func Middleware(ab *Authboss, redirectToLogin bool, allowHalfAuth bool, force2fa
 				if redirectToLogin {
 					log.Infof("redirecting unauthorized user to login from: %s", r.URL.Path)
 					vals := make(url.Values)
-					vals.Set(FormValueRedirect, r.URL.Path)
+
+					redirURL := r.URL.Path
+					if mountPathed && len(ab.Config.Paths.Mount) != 0 {
+						redirURL = path.Join(ab.Config.Paths.Mount, redirURL)
+					}
+					vals.Set(FormValueRedirect, redirURL)
 
 					ro := RedirectOptions{
 						Code:         http.StatusTemporaryRedirect,
@@ -108,15 +129,15 @@ func Middleware(ab *Authboss, redirectToLogin bool, allowHalfAuth bool, force2fa
 
 					if err := ab.Config.Core.Redirector.Redirect(w, r, ro); err != nil {
 						log.Errorf("failed to redirect user during authboss.Middleware redirect: %+v", err)
-						return
 					}
+					return
 				}
 
 				log.Infof("not found for unauthorized user at: %s", r.URL.Path)
 				w.WriteHeader(http.StatusNotFound)
 			}
 
-			if !allowHalfAuth && !IsFullyAuthed(r) || !force2fa && !IsTwoFactored(r) {
+			if forceFullAuth && !IsFullyAuthed(r) || force2fa && !IsTwoFactored(r) {
 				fail(w, r)
 				return
 			}